# api/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Simple Movies list API",
    "description" : "This is a simple API for managing movies and comments",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/bopzor/wat-wat/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "paths" : {
    "/movies" : {
      "get" : {
        "tags" : [ "movies" ],
        "summary" : "get the movies list",
        "operationId" : "getMoviesList",
        "responses" : {
          "200" : {
            "description" : "movies list fetched",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "movies" ],
        "summary" : "adds movie",
        "operationId" : "createMovie",
        "requestBody" : {
          "description" : "Movie to add",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MoviePartial"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "created movie",
            "content" : {
              "appliation/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}" : {
      "put" : {
        "tags" : [ "movies" ],
        "summary" : "updates an existing movie",
        "operationId" : "updateMovie",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of movie to update",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "description" : "Movie to update",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "updated movie",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "movies" ],
        "summary" : "delete movie matching given id",
        "operationId" : "deleteMovie",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of movie to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "no content"
          }
        }
      }
    },
    "/movies/sort" : {
      "post" : {
        "tags" : [ "movies" ],
        "summary" : "reorder movies",
        "operationId" : "sortMovies",
        "requestBody" : {
          "description" : "Movie to add",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Order"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "return the new sorted movies list",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}/comments" : {
      "post" : {
        "tags" : [ "comments" ],
        "summary" : "add a comment in an existing movie",
        "operationId" : "createComment",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of movie where to add the comment",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "description" : "Comment to add",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CommentPartial"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "updated movie",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}/comments/{commentId}" : {
      "put" : {
        "tags" : [ "comments" ],
        "summary" : "update an existing comment",
        "operationId" : "updateComment",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of movie where to remove the comment",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "description" : "ID of the comment to remove",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "description" : "Comment to update",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Comment"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "updated movie",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "comments" ],
        "summary" : "remove an existing",
        "operationId" : "deleteComment",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of movie where to remove the comment",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "description" : "ID of the comment to remove",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "no content"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Movie" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 42
          },
          "title" : {
            "type" : "string",
            "example" : "Toy Story"
          },
          "plot" : {
            "type" : "string",
            "example" : "A cowboy doll is profoundly threatened and jealous when a new spaceman figure supplants him as top toy in a boy's room."
          },
          "released" : {
            "type" : "string",
            "example" : "22 Nov 1995"
          },
          "runtime" : {
            "type" : "string",
            "example" : "81 min"
          },
          "director" : {
            "type" : "string",
            "example" : "John Lasseter"
          },
          "writer" : {
            "type" : "string",
            "example" : "John Lasseter (original story by), Pete Docter (original story by), Andrew Stanton (original story by), Joe Ranft (original story by), Joss Whedon (screenplay by), Andrew Stanton (screenplay by), Joel Cohen (screenplay by), Alec Sokolow (screenplay by)"
          },
          "actors" : {
            "type" : "string",
            "example" : "Tom Hanks, Tim Allen, Don Rickles, Jim Varney"
          },
          "poster" : {
            "type" : "string",
            "example" : "ttps://m.media-amazon.com/images/M/MV5BMDU2ZWJlMjktMTRhMy00ZTA5LWEzNDgtYmNmZTEwZTViZWJkXkEyXkFqcGdeQXVyNDQ2OTk4MzI@._V1_SX300.jpg"
          },
          "imdbID" : {
            "type" : "string",
            "example" : "tt0114709"
          },
          "seen" : {
            "type" : "boolean",
            "default" : false
          },
          "place" : {
            "type" : "integer",
            "example" : 12
          },
          "createdAt" : {
            "type" : "string",
            "example" : "2019-03-03T19:23:50.848Z"
          },
          "updatedAt" : {
            "type" : "string",
            "example" : "2019-03-04T19:23:50.848Z"
          },
          "deletedAt" : {
            "type" : "string"
          }
        }
      },
      "MoviePartial" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "example" : "Toy Story"
          },
          "plot" : {
            "type" : "string",
            "example" : "A cowboy doll is profoundly threatened and jealous when a new spaceman figure supplants him as top toy in a boy's room."
          },
          "released" : {
            "type" : "string",
            "example" : "22 Nov 1995"
          },
          "runtime" : {
            "type" : "string",
            "example" : "81 min"
          },
          "director" : {
            "type" : "string",
            "example" : "John Lasseter"
          },
          "writer" : {
            "type" : "string",
            "example" : "John Lasseter (original story by), Pete Docter (original story by), Andrew Stanton (original story by), Joe Ranft (original story by), Joss Whedon (screenplay by), Andrew Stanton (screenplay by), Joel Cohen (screenplay by), Alec Sokolow (screenplay by)"
          },
          "actors" : {
            "type" : "string",
            "example" : "Tom Hanks, Tim Allen, Don Rickles, Jim Varney"
          },
          "poster" : {
            "type" : "string",
            "example" : "ttps://m.media-amazon.com/images/M/MV5BMDU2ZWJlMjktMTRhMy00ZTA5LWEzNDgtYmNmZTEwZTViZWJkXkEyXkFqcGdeQXVyNDQ2OTk4MzI@._V1_SX300.jpg"
          },
          "imdbID" : {
            "type" : "string",
            "example" : "tt0114709"
          }
        }
      },
      "Order" : {
        "required" : [ "order" ],
        "type" : "object",
        "properties" : {
          "order" : {
            "$ref" : "#/components/schemas/Order_order"
          }
        }
      },
      "CommentPartial" : {
        "type" : "object",
        "properties" : {
          "author" : {
            "type" : "string",
            "example" : "Toto"
          },
          "comment" : {
            "type" : "string",
            "example" : "I love this movie I really want to see it again."
          }
        }
      },
      "Comment" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 6
          },
          "comment" : {
            "type" : "string",
            "example" : "I love this movie I really want to see it again."
          },
          "author" : {
            "type" : "string",
            "example" : "Toto"
          },
          "createdAt" : {
            "type" : "string",
            "example" : "2018-12-13T16:22:23.376Z"
          },
          "updatedAt" : {
            "type" : "string",
            "example" : "2018-12-13T16:22:45.645Z"
          },
          "deletedAt" : {
            "type" : "string"
          },
          "movieId" : {
            "type" : "integer",
            "example" : 42
          }
        }
      },
      "Order_order" : {
        "type" : "object",
        "properties" : {
          "prev" : {
            "type" : "integer"
          },
          "nxt" : {
            "type" : "integer"
          }
        }
      }
    }
  }
}